job, parallel: add support for failfast tests when running with parallel
When runnign tests with parallel, if a failfast test exits with an error
kill all the other tests.

If the subprocess is killed it will return rc=15 for SIGTERM.
fork_poll will always raise an exception if the rc is not 0, so checking
for rc > 0 is pointless

We had to add logic to fork_start to pass return codes to os._exit().
Most things will return None, but job.groupfunc or something will return
True/False, so we have to negate the True/False for os._exit()

Refactored the task killing, if we get a failfast we just loop once and
poll, for things that aren't finished we kill right there.

Added PID tracking for most cases.

Set attributes 'failfast', 'ignore_status' or
'last' per task.  'last' means we kill all the other tasks upon exit,
even if there is no error.  'ignore_status' marks jobs that don't get a
poll after the kill so they don't pick up the rc=15 SIGTERM exit status.
This is need for things like dmesg handling where we don't really want
to know the exit status unless it fails fast.

TODO: add support for WIFSIGNALED so we can tell when processes where
killed with a signal, versus other deaths.

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>